src/text-harvester/requirements.txt
yt-dlp
SpeechRecognition
python-dotenv